+++ src/Components/BreadCrumbs/BreadCrumbs.tsx
@@ -1,8 +1,6 @@
 import { Link } from "react-router-dom";
 import { Icon } from "@justfixnyc/component-library";
 import "./BreadCrumbs.scss";
-import { Address } from "../Pages/Home/Home";
-import { toTitleCase } from "../../helpers";
 
 interface BreadCrumbs {
   crumbs: { path?: string; name?: string; active?: boolean }[];
@@ -36,13 +34,3 @@ export const BreadCrumbs: React.FC<BreadCrumbs> = ({ crumbs }) => {
   });
   return <div className="breadCrumbs">{crumbpath}</div>;
 };
-
-export function breadCrumbAddress(address?: Address, maxLength = 45) {
-  if (!address) return "Your address";
-  const addrShort = toTitleCase(
-    `${address?.houseNumber} ${address?.streetName}`
-  );
-  return addrShort.length > maxLength
-    ? `${addrShort.substring(0, maxLength)}\u2026`
-    : addrShort;
-}
