+++ src/Components/Pages/Form/Form.tsx
@@ -7,9 +7,9 @@ import { Address } from "../Home/Home";
 import { useGetBuildingData, useSendGceData } from "../../../api/hooks";
 import { useSessionStorage } from "../../../hooks/useSessionStorage";
 import { RadioGroup } from "../../RadioGroup/RadioGroup";
-import { breadCrumbAddress, BreadCrumbs } from "../../BreadCrumbs/BreadCrumbs";
+import { BreadCrumbs } from "../../BreadCrumbs/BreadCrumbs";
 import { InfoBox } from "../../InfoBox/InfoBox";
-import { formatNumber } from "../../../helpers";
+import { breadCrumbAddress, formatNumber } from "../../../helpers";
 import { cleanFormFields } from "../../../api/helpers";
 import { BuildingData, GCEUser } from "../../../types/APIDataTypes";
 import "./Form.scss";
