+++ src/helpers.ts
@@ -5,6 +5,7 @@ import {
   CriteriaDetermination,
   WowBuildings,
 } from "./types/APIDataTypes";
+import { Address } from "./Components/Pages/Home/Home";
 
 export function toTitleCase(x: string) {
   return x.replace(
@@ -22,6 +23,16 @@ export const formatGeosearchAddress = (
         properties.postalcode
       }`
     : "";
+
+export function breadCrumbAddress(address?: Address, maxLength = 45) {
+  if (!address) return "Your address";
+  const addrShort = toTitleCase(
+    `${address?.houseNumber} ${address?.streetName}`
+  );
+  return addrShort.length > maxLength
+    ? `${addrShort.substring(0, maxLength)}\u2026`
+    : addrShort;
+}
 
 export const getDetermination = (
   eligibilityResults?: EligibilityResults
